updated tmp creation

+++ manifests/init.pp
@@ -37,49 +37,52 @@ class mountebs (
     ensure => present
   }
 
-  mdadm { '/dev/md0' :
-    ensure    => 'created',
-    devices   => ['/dev/xvdb', '/dev/xvdc'],
-    level     => 0,
-    force     => true,
-    notify    => Exec['format /dev/md0'],
-    onlyif    => $set_tmp_dir
+  if ($set_tmp_dir) {
+    mdadm { '/dev/md0' :
+      ensure    => 'created',
+      devices   => ['/dev/xvdb', '/dev/xvdc'],
+      level     => 0,
+      force     => true,
+      notify    => Exec['format /dev/md0'],
+      onlyif    => $set_tmp_dir
+    }
+
+    exec {'format /dev/md0':
+      command     => 'mkfs.ext4 -j -F /dev/md0',
+      path        => '/sbin',
+      refreshonly => true,
+      notify      => Exec['label /tmp'],
+      onlyif      => $set_tmp_dir
+    }
+
+    file {'set tmp mount point':
+      path     => '/tmp',
+      ensure   => directory,
+      mode     => 'ug=rwx,o=rwxt',
+      onlyif   => $set_tmp_dir
+    }
+
+    # potrebbe esserci un problema se non fa il label prima del mount. Controllare perché aggiunto il refreshonly
+    exec {'label /tmp':
+      command     => "e2label /dev/md0 instance_store",
+      path        => "/sbin",
+      refreshonly => true,
+      onlyif      => $set_tmp_dir
+    }
+
+    mount {'/tmp':
+      ensure    => 'mounted',
+      atboot    => true,
+      device    => 'LABEL=instance_store',
+      fstype    => 'auto',
+      options   => 'defaults',
+      onlyif    => $set_tmp_dir
+    }
+    
+    Exec['umount cmd for mnt'] -> Mount['umount /mnt'] -> File['/mnt/apps'] -> Mount['/mnt/apps'] -> File['/home/ubuntu/current']
+    
+  }else{
+    Exec['umount cmd for mnt'] -> Mount['umount /mnt'] -> File['/mnt/apps'] -> Mount['/mnt/apps'] -> File['/home/ubuntu/current'] -> Package['mdadm'] -> Mdadm['/dev/md0'] -> Mount['/tmp'] -> File['set tmp mount point']
   }
-
-  exec {'format /dev/md0':
-    command     => 'mkfs.ext4 -j -F /dev/md0',
-    path        => '/sbin',
-    refreshonly => true,
-    notify      => Exec['label /tmp'],
-    onlyif      => $set_tmp_dir
-  }
-
-  file {'set tmp mount point':
-    path     => '/tmp',
-    ensure   => directory,
-    mode     => 'ug=rwx,o=rwxt',
-    onlyif   => $set_tmp_dir
-  }
-
-  # potrebbe esserci un problema se non fa il label prima del mount. Controllare perché aggiunto il refreshonly
-  exec {'label /tmp':
-    command     => "e2label /dev/md0 instance_store",
-    path        => "/sbin",
-    refreshonly => true,
-    onlyif      => $set_tmp_dir
-  }
-
-  mount {'/tmp':
-    ensure    => 'mounted',
-    atboot    => true,
-    device    => 'LABEL=instance_store',
-    fstype    => 'auto',
-    options   => 'defaults',
-    onlyif    => $set_tmp_dir
-  }
-
-
-   Exec['umount cmd for mnt'] -> Mount['umount /mnt'] -> File['/mnt/apps'] -> Mount['/mnt/apps'] -> File['/home/ubuntu/current'] -> Package['mdadm'] -> Mdadm['/dev/md0'] -> Mount['/tmp'] -> File['set tmp mount point']
-
 
 }
